Improved caching for snapshot lookups
